Extract canonical identifier normalizers to sbir_etl/utils/identifiers.py
normalize_uei, normalize_duns, and normalize_cage were each defined in
multiple places (company.py, organization.py, phase_transition/utils.py)
with minor variations in NA handling and calling convention.

New sbir_etl/utils/identifiers.py provides one authoritative implementation
of each, backed by _blank() for consistent NA handling.

- company.py: validate_duns/cage now delegate to normalize_duns/cage and
  raise ValueError on bad format (preserved existing strictness)
- organization.py: same for duns/cage; validate_uei delegates to
  normalize_uei and stays lenient (returns None, no raise)
- phase_transition/utils.py: local definitions removed; functions imported
  from identifiers.py and re-exported via __all__ so phase_ii/phase_iii
  callers are unaffected
